refactor: move useSave and useMenu to lib
not much interaction with budget related stuff

also remove useBudgetState abstraction in favor of direct use of
budgetReducer
