+++ src/budget/index.ts
@@ -12,10 +12,8 @@ import { Action as ActionT } from './budgetReducer';
 
 export * from './budgetReducer';
 export { VERSION } from './Types';
-export { default } from './useBudgetState';
 export { default as useBudgetData } from './useBudgetData';
 export { default as getInitData, INIT_EMPTY } from './getInitData';
-export { unsaved } from './useSave';
 
 export type Action = ActionT;
 export type BudgetState = BudgetStateT;
