+++ src/views/Budget/Entry.tsx
@@ -1,7 +1,11 @@
-import React, { lazy } from 'react';
-import { Startup } from '../../components';
-import useBudgetState, { BudgetState } from '../../budget';
-import { withShowSettingsProvider, useShowSettings } from '../../lib';
+import React, { lazy, useReducer } from 'react';
+import { BudgetState, budgetReducer } from '../../budget';
+import {
+  withShowSettingsProvider,
+  useShowSettings,
+  useMenu,
+  useSave,
+} from '../../lib';
 import Settings from '../Settings';
 
 type Props = {
@@ -11,17 +15,18 @@ type Props = {
 const Budget = lazy(() => import('./Budget'));
 
 export default withShowSettingsProvider(({ initialState }: Props) => {
+  const [state, dispatch] = useReducer(budgetReducer, initialState);
+  const menu = useMenu();
+  const error = useSave(menu, state);
   const showSettings = useShowSettings();
-  const { error, state, dispatch } = useBudgetState(initialState);
+
   if (error) {
     throw error;
   }
-  if (state === null) {
-    return <Startup />;
-  }
-  if (showSettings) {
-    return <Settings state={state} dispatch={dispatch} />;
-  }
 
-  return <Budget state={state} dispatch={dispatch} />;
+  return showSettings ? (
+    <Settings state={state} dispatch={dispatch} />
+  ) : (
+    <Budget state={state} dispatch={dispatch} />
+  );
 });
