+++ src/views/NewBudget/NewBudget.tsx
@@ -1,14 +1,14 @@
 import React, { useState, useReducer } from 'react';
 import { remote } from 'electron';
-import { BudgetState, VERSION, unsaved } from '../../budget';
+import { BudgetState, VERSION } from '../../budget';
 import { Content, Button, Header, HeaderSpacer } from '../../components';
 import startOfMonth from 'date-fns/startOfMonth';
 import subMonths from 'date-fns/subMonths';
 import General from '../Settings/General';
 import Categories from '../Settings/Categories';
-import budgetReducer from '../../budget/budgetReducer';
-import useMenu from '../../budget/useMenu';
-import { getToday } from '../../lib';
+import { budgetReducer } from '../../budget';
+import useMenu from '../../lib/useMenu';
+import { getToday, unsaved } from '../../lib';
 
 type Props = {
   onCreate: (budget: BudgetState) => void;
